Move redis_file_storable.rb into export.rb
